Scraping/requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
schedule==1.2.1
SQLAlchemy==2.0.23
mysqlclient==2.2.0
python-dotenv==1.0.0
